Ensure that `--sandbox` affects `--embed-resources`.
Previously it did not (contrary to what was implied by the manual),
which means that an image with URL `/etc/passwd` would leak an
encoded version of that file to HTML output with `--self-contained`
or `--embed-resources`, even if `--sandbox` was used.

Thanks to Samuel Mortenson for pointing out the issue.