ODT reader: fix unwanted block quotes.
Previously the reader created block quotes whenever a paragraph
was marked indented (even though this just affects the first line).
With this change we still generate block quotes for content that
has an altered left margin, but not for indented paragraphs.

See #10575.

This patch does NOT address the related #9505 which concerns lists.